+++ .eslintrc.json
@@ -22,6 +22,7 @@
       "linebreak-style": 0,
       "max-len": 0,
       "react/jsx-filename-extension": 0,
-      "no-alert": 0
+      "no-alert": 0,
+      "react/prop-types": 0
     }
 }
