+++ src/components/home.jsx
@@ -20,7 +20,6 @@ const retrieveUser = async (dbRef) => {
     isAdmin: temp.data().isAdmin,
     prevShifts: ps,
   };
-  console.log(user);
 
   return user;
 };
@@ -82,15 +81,9 @@ export default function Home() {
       const [user, users, vigils] = firestoreResponse;
 
       // Initialize redux store
-      console.log('Home: Initialize User store');
       dispatch(actions.user.initializeUser(user));
-
-      console.log('Home: Initialize Users store');
       dispatch(actions.users.initializeUsers(users));
-
-      console.log('Home: Initialize Vigils store');
       dispatch(actions.vigils.initalizeVigils(vigils));
-
       // navigation.navigate('Root');
     };
