+++ src/components/pages/discussion/Discussion.jsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import Row from 'react-bootstrap/Row';
 import Container from 'react-bootstrap/Container';
 import { Link } from 'react-router-dom';
-import HeaderWithNav from '../../navigation/nav-header';
 import { useSelector } from 'react-redux';
+import HeaderWithNav from '../../navigation/nav-header';
 import CreateThread from './createThread';
 import * as constants from '../../../constants';
 import Forum from './Forum';
