+++ src/components/pages/schedule/Schedule.jsx
@@ -1,121 +1,55 @@
 import React, { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
-import Modal from 'react-bootstrap/Modal';
-import Button from 'react-bootstrap/Button';
 import styled from 'styled-components';
 import { useSelector } from 'react-redux';
-import firebase from 'firebase';
 import HeaderWithNav from '../../navigation/nav-header';
-import ShiftDetails from './shiftDetails';
-import Calendar from './Calendar';
-import eventPropType from '../../../dataStructures/propTypes';
+import Calendar from './calendar/Calendar';
 
-const StyledButton = styled.button`
-  color: white;
-  background-color: #84C0C9;
-  border: 2px solid #84C0C9;
-  border-radius: 5px;
-  width: 100%;
-  padding: 6px 0px; 
-  font-size: 14px;
-  fontFamily: Roboto;
-  
-  &:hover{
-    color: white;
-    background-color: #558E97;
-  }
-`;
 const PaddedDiv = styled.div`
   padding: 0 2%;
 `;
-async function addShiftPress() {
-  // creates a new shift and adds it to a specific vigil
 
-  console.log('press');
-  const currentUser = firebase.auth().currentUser.uid;
-  //  console.log("ID " + vigil.id);
-  const db = firebase.firestore();
-  const vigilRef = db.collection('vigils').doc('kEtigasg0zFzkhYBaWGc');
-  vigilRef.collection('shift').add({
-    shiftStartTime: 'start', // vigil.startTime,
-    shiftEndTime: 'end', // vigil.endTime,
-    userID: currentUser,
-  })
-    .then(() => {
-      console.log('Document successfully written!');
-    })
-    .catch((error) => {
-      console.error('Error writing document: ', error);
-    });
-}
-
-export default function Schedule({ selectVigil, setSelectVigil }) {
-  const [show, setShow] = useState(false);
-  const [vigils, setVigils] = useState([]);
+export default function Schedule(props) {
+  const { setSelectVigil } = props;
+  const [eventData, setEventData] = useState([]);
 
   // Gets Vigil Data from redux store
   const storeVigils = useSelector((state) => state.vigils.vigils);
-  console.log(storeVigils);
+
+  // Come back to this based off of desired
+  // start/end time format. This isnt set up for recurring,
+  // If this is desired functionalilty. swap array structure for a start/end timestamp structure.
+  const getVigilInfo = () => {
+    const vigilsData = [];
+    storeVigils.forEach((v) => {
+      vigilsData.push({
+        title: v.address,
+        start: `${v.dates[0]}T${v.startTime}`,
+        end: `${v.dates[v.dates.length - 1]}T${v.endTime}`,
+        backgroundColor: (v.dates.length !== 1) ? '#8FCBD4' : '#D0CAEB', // if single/multi day event
+        notes: v.notes,
+        id: v.id,
+        dates: v.dates,
+        sTime: v.startTime,
+        eTime: v.endTime,
+      });
+    });
+    setEventData(vigilsData);
+  };
 
   const fetchData = async () => {
-    setVigils(storeVigils);
+    await getVigilInfo();
   };
 
   useEffect(() => {
     fetchData();
   }, []); // This useEffect block gets whole collection of vigil documents upon rendering of this component
 
-  function handleClose() {
-    setShow(false);
-    fetchData(); // After delete, we now want to refetch the newly updated document collection, this function is passed down as prop to shiftDetails
-  }
-
-  function handleShow(vigil) {
-    setShow(true);
-    setSelectVigil((prevState) => ({
-      ...prevState,
-      id: vigil.id,
-      address: vigil.address,
-      dates: vigil.dates,
-      startTime: vigil.startTime,
-      endTime: vigil.endTime,
-      notes: vigil.notes,
-    }));
-  }
-
   return (
     <div>
       <HeaderWithNav>Schedule</HeaderWithNav>
       <PaddedDiv>
-        {vigils.map((vigil) => (
-          <div key={vigil.id}>
-            <Button variant="primary" size="sm" onClick={() => handleShow(vigil)}>{vigil.address}</Button>
-          </div>
-        ))}
-        <Modal show={show} size="lg" onEscapeKeyDown={handleClose} onHide={handleClose} centered>
-          <Modal.Header closeButton>Shift Details</Modal.Header>
-          <Modal.Body>
-            <ShiftDetails
-              func={handleClose}
-              id={selectVigil.id}
-              address={selectVigil.address}
-              dates={selectVigil.dates}
-              startTime={selectVigil.startTime}
-              endTime={selectVigil.endTime}
-              notes={selectVigil.notes}
-            />
-          </Modal.Body>
-          <Modal.Footer>
-            <StyledButton onClick={addShiftPress}>Sign Up</StyledButton>
-          </Modal.Footer>
-        </Modal>
-        <Calendar />
+        <Calendar eventData={eventData} setSelectVigil={setSelectVigil} />
       </PaddedDiv>
     </div>
   );
 }
-
-Schedule.propTypes = {
-  selectVigil: eventPropType.isRequired,
-  setSelectVigil: PropTypes.func.isRequired,
-};
